Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Thanks Remi for test & rls submission

No problem. Btw, I guess we want to keep the bootstrap blueprints up to date, but release and merge that into OOM at the end of Istanbul? (now only dcaegen2-services update will suffice?)

Also the merge job seems to be broken, I think it's because the maven-deploy-plugin:2.8.2 is incompatibile with mvn33, so we could update the jenkins job to use mvn35 (preferred probably) or change the plugin to version 2.8.
Created a jira for that: https://jira.onap.org/browse/DCAEGEN2-2797 , I'll probably look into it tomorrow or next week.

Patch-set: 1